moved openTunnel to a separate module
