+++ package.json
@@ -48,6 +48,7 @@
     "chai": "^3.5.0",
     "chai-as-promised": "^6.0.0",
     "isparta": "^4.0.0",
-    "mocha": "^3.2.0"
+    "mocha": "^3.2.0",
+    "sinon": "^1.17.6"
   }
 }
