+++ test/goodActivity.js
@@ -4,7 +4,7 @@ export default class MockActivity extends Activity {
 
     async onTask(task, config, executionDate) {
         return await new Promise((resolve, reject) => {
-            setTimeout(resolve.bind(null, 'result'), 1000);
+            resolve('result');
         });
     }
 };
